refactor: extract videoGen POST param preparation into a service (#3288)
The POST / handler carried ~645 lines of validation, durable upload staging
with rollback bookkeeping, IC-LoRA/keyframe/a2v gating, render-history
resolution and LoRA/chunk normalization inline — twenty-plus ServerError
throw sites sharing mutable cleanup state. Per the module contract that
belongs in server/services/, and the sibling imageGen route already does
exactly this via services/imageGen/prepareParams.js.

Behavior is preserved verbatim: the validation sequence is load-bearing
(several throws depend on an earlier one having run) and every early return
still unwinds the staged uploads in the same order. The existing 105-case
route suite passes unchanged; a new service suite pins the contract at the
new boundary.

Also closes the rejected-await leak (#3326) the extraction made natural to
fix: only the explicit throws called cleanupAllStaged(), so a rejected
getMusicVideoProject/getTrack/loadHistory/enqueueJob bubbled past them and
orphaned every durable copy under data/uploads (the job never exists, so the
worker never cleans up either). withStagedRollback wraps the whole staging
region plus the enqueue, releasing the copies and rethrowing the original
error untouched.

Closes #3288
Closes #3326
